add show tasks at picked date
