+++ components/organisms/feed/feed.tsx
@@ -1,112 +1,43 @@
 import {Book as BookOpen, BookHalf} from "@styled-icons/bootstrap";
 import {Book as BookClose, BookDead} from "@styled-icons/fa-solid";
-import React, {useState, useEffect, useCallback,useMemo} from "react";
+import React, {useMemo} from "react";
 import * as S from "./feed.styled";
 import {TaskList} from "./task-list";
 import {Chip} from "components/atoms/chip";
 import {ReducedGoalData} from "pages";
-import {useDataSaga, DataActionType, DataSagaStatus,TaskData} from "stores/data";
+import {TaskData, GoalData} from "stores/data";
 import {Dayjs} from "utils/dayjs";
 
 type FeedProps = {
   pickedDate?:string,
-	goals: ReducedGoalData[];
+  goalTasks: Record<string, TaskData[]>,
+	goals: ReducedGoalData[],
+  onTaskDelete: (id:string)=>void,
+  onTaskCreate: (id:string)=>void,
+  onTaskUpdate:(id:string, text:string)=>void,
 };
 
 export const Feed = ({
   pickedDate,
-  goals
+  goalTasks,
+  goals,
+  onTaskDelete,
+  onTaskCreate,
+  onTaskUpdate
 }: FeedProps) => {
-  const {
-    fetch: getTasksByDaysFetch,
-    data: getTasksByDaysData,
-    refetch: getTasksByDaysRefetch,
-  } = useDataSaga<DataActionType.GET_TASKS_BY_DAYS>(DataActionType.GET_TASKS_BY_DAYS);
-  const {
-    fetch: createTaskFetch, 
-    status: createTaskStatus
-  } = useDataSaga<DataActionType.CREATE_TASK>(DataActionType.CREATE_TASK);
-  const {
-    fetch: deleteTaskFetch, 
-    status: deleteTaskStatus
-  } = useDataSaga<DataActionType.DELETE_TASK>(DataActionType.DELETE_TASK);
-  const {
-    fetch: updateTaskFetch, 
-    status: updateTaskStatus
-  } = useDataSaga<DataActionType.UPDATE_TASK>(DataActionType.UPDATE_TASK);
-  
-  const [tasks, setTasks] = useState<TaskData[]>(getTasksByDaysData || []);
+  const formatedDate = useMemo(()=>{
+    const year = pickedDate?.slice(4);
+    const month =  pickedDate?.slice(2,4)
+    const date = pickedDate?.slice(0,2);
 
-  const handleTaskCreate = useCallback(
-    (id: string) => {
-      createTaskFetch({
-        data: {
-          title: "",
-          goal:id,
-          doneAt: new Date().getTime(),
-        },
-      });
-    },
-    [createTaskFetch]
-  );
-
-  const handleTaskDelete = useCallback((id: string)=>{
-    deleteTaskFetch({
-      pathSegments: [id]
-    })
-  },[deleteTaskFetch])
-
-  const handleTaskUpdate = useCallback((id:string,title:string)=>{
-    updateTaskFetch({
-      pathSegments: [id],
-      data: {
-        title,
-      }});
-  },[updateTaskFetch])
-  
-  const goalTasks = useMemo(()=>{
-    const newGoalTasks: Record<string, TaskData[]> = {};
-    goals.forEach(goal=>{
-      newGoalTasks[goal.id] = tasks.filter(taskItem => 
-        taskItem.goal === goal.id && 
-        Dayjs(taskItem.doneAt).format("DDMMYYYY") === pickedDate)
-    })
-    
-    return newGoalTasks;
-  },[goals,tasks,pickedDate])
-
-  useEffect(() => {
-    getTasksByDaysFetch({
-      startDay: new Date("1999-11-11"),
-      endDay: new Date("2222-11-11"),
-    });
-  }, [getTasksByDaysFetch]);
-
-  useEffect(() => {
-    setTasks(getTasksByDaysData || []);
-  }, [getTasksByDaysData]);
-  
-  useEffect(()=>{
-    if (createTaskStatus === DataSagaStatus.SUCCEEDED){
-      getTasksByDaysRefetch()
-    }
-  },[getTasksByDaysRefetch, createTaskStatus])
-  
-  useEffect(()=>{
-    if (deleteTaskStatus === DataSagaStatus.SUCCEEDED){
-      getTasksByDaysRefetch()
-    }
-  },[getTasksByDaysRefetch, deleteTaskStatus])
-
-  useEffect(() => {
-    if (updateTaskStatus === DataSagaStatus.SUCCEEDED) {
-      getTasksByDaysRefetch();
-    }
-  }, [getTasksByDaysRefetch, updateTaskStatus]);
+    return `${year}년 ${month}월 ${date}일`;
+  },[pickedDate]);
 
   return (
     <S.Feed>
-      <S.Header>Feed</S.Header>
+      <S.Header>Feed
+        <S.PickedDate>{formatedDate}</S.PickedDate>
+      </S.Header>
       <S.FeedContents>
         {goals.map((goal) => (
           <S.GoalAndInput key={goal.id}>
@@ -115,13 +46,13 @@ export const Feed = ({
               label={goal.name}
               color={goal.color}
               icon={<BookClose />}
-              onAdd={handleTaskCreate}
+              onAdd={onTaskCreate}
             ></Chip>
             <TaskList
               color={goal.color}
               tasks={goalTasks[goal.id]}
-              onTaskDelete={handleTaskDelete}
-              onTaskUpdate={handleTaskUpdate}
+              onTaskDelete={onTaskDelete}
+              onTaskUpdate={onTaskUpdate}
             ></TaskList>
           </S.GoalAndInput>
         ))}
